backend/requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
httpx>=0.28.0

# Celery & Redis
celery[redis]>=5.4.0
redis>=5.2.0

# FFmpeg
static-ffmpeg>=2.7

# AI Translation
openai>=1.50.0
google-genai>=1.0.0
deepl>=1.18.0

# Subtitle parsing
pysubs2>=1.7.0
chardet>=5.2.0

# Utils
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
aiofiles>=24.1.0
tenacity>=9.0.0
structlog>=24.4.0
PyJWT>=2.9.0
